move utility functions to dedicated file
